backend/requirements.txt
fastapi==0.109.0
uvicorn==0.30.0
sqlalchemy==2.0.40
pydantic==2.11.2
pydantic-settings==2.0.3
pydantic[email]==2.11.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.11
alembic==1.15.2
psycopg2-binary==2.9.9
bcrypt==4.0.1
python-jose==3.3.0
